# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GoGym API",
        "description": "API REST para gerenciamento de academias, personal trainers e alunos. Permite registro de usuários (personal e aluno), autenticação via token (Sanctum), verificação de e-mail, CRUD completo de personais e alunos, e sistema de convites para vinculação entre personal e aluno.",
        "contact": {
            "name": "Gustavo",
            "email": "nao-responda@ugustavodev.com.br"
        },
        "license": {
            "name": "Proprietária",
            "url": "https://ugustavodev.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:84/api",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
        "/aluno": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Listar alunos",
                "description": "Retorna uma lista paginada de alunos. Se o usuário autenticado for um personal, retorna apenas seus alunos vinculados. Se for um aluno, retorna apenas seus próprios dados. Requer autenticação e e-mail verificado.",
                "operationId": "listarAlunos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de alunos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AlunoResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginacaoLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginacaoMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "E-mail não verificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutorizado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/aluno/{aluno}": {
            "get": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Exibir um aluno específico",
                "description": "Retorna os dados detalhados de um aluno pelo seu ID. O acesso é controlado por policy (apenas o próprio aluno ou o personal vinculado pode visualizar). Requer autenticação e e-mail verificado.",
                "operationId": "exibirAluno",
                "parameters": [
                    {
                        "name": "aluno",
                        "in": "path",
                        "description": "ID do aluno",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do aluno retornados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlunoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutorizado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aluno não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Atualizar dados de um aluno",
                "description": "Atualiza parcialmente os dados de um aluno. Todos os campos são opcionais. Autorização é controlada por policy. Requer autenticação e e-mail verificado.",
                "operationId": "atualizarAluno",
                "parameters": [
                    {
                        "name": "aluno",
                        "in": "path",
                        "description": "ID do aluno",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos a serem atualizados",
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telefone": {
                                        "type": "string",
                                        "example": "(11) 96666-6666",
                                        "maxLength": 20
                                    },
                                    "genero": {
                                        "type": "string",
                                        "example": "feminino",
                                        "nullable": true,
                                        "enum": [
                                            "masculino",
                                            "feminino",
                                            "nao_binario",
                                            "outro",
                                            "prefiro_nao_informar"
                                        ]
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1998-05-20",
                                        "nullable": true
                                    },
                                    "peso": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 70,
                                        "nullable": true
                                    },
                                    "altura": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1.75,
                                        "nullable": true
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Aluno atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AlunoResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutorizado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aluno não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroValidacao"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Alunos"
                ],
                "summary": "Inativar um aluno",
                "description": "Marca o aluno como inativo e realiza soft delete. Requer autenticação e e-mail verificado.",
                "operationId": "deletarAluno",
                "parameters": [
                    {
                        "name": "aluno",
                        "in": "path",
                        "description": "ID do aluno",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aluno inativado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Aluno inativado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoAutenticado"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aluno não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErroNaoEncontrado"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErroValidacao": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "ErroNaoAutenticado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ErroNaoAutorizado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "This action is unauthorized."
                    }
                },
                "type": "object"
            },
            "ErroNaoEncontrado": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Recurso não encontrado."
                    }
                },
                "type": "object"
            },
...
            "AlunoResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "João Silva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@email.com"
                    },
                    "telefone": {
                        "type": "string",
                        "example": "(11) 99999-9999"
                    },
                    "genero": {
                        "type": "string",
                        "example": "masculino",
                        "nullable": true,
                        "enum": [
                            "masculino",
                            "feminino",
                            "nao_binario",
                            "outro",
                            "prefiro_nao_informar"
                        ]
                    },
                    "dados_fisicos": {
                        "properties": {
                            "peso": {
                                "type": "number",
                                "format": "float",
                                "example": 75.5,
                                "nullable": true
                            },
                            "altura": {
                                "type": "number",
                                "format": "float",
                                "example": 1.78,
                                "nullable": true
                            },
                            "data_nascimento": {
                                "type": "string",
                                "format": "date",
                                "example": "1998-05-20",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "status_vinculo": {
                        "type": "string",
                        "example": "ativo",
                        "nullable": true
                    },
                    "status_conta": {
                        "type": "string",
                        "example": "Ativo",
                        "enum": [
                            "Ativo",
                            "Inativo"
                        ]
                    },
                    "cadastrado_em": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-13 22:00:00"
                    }
                },
                "type": "object"
            },
            "PaginacaoLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginacaoMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 5
                    },
                    "total": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Insira o token retornado no login/registro. Exemplo: {token}",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
...
}